+++ internal/meraki/devices/cellular/resource_schema.go
