+++ internal/meraki/devices/switch/ports/datasource_schema.go
